requirements/production.txt
Django==1.9.1
python-social-auth
django-pipeline
pillow
django-summernote
djangorestframework
bs4
django-crispy-forms
django-mptt
